requirements.txt
streamlit
python-dotenv
google-genai
pymupdf
chromadb
sentence-transformers